@@ -11,10 +11,8 @@ use Twig\TwigFunction;
 
 class FMElfinderExtension extends AbstractExtension
 {
-    /**
-     * @var Environment
-     */
-    protected $twig;
+
+    protected Environment $twig;
 
     public function __construct(Environment $twig)
     {
@@ -33,21 +31,19 @@ class FMElfinderExtension extends AbstractExtension
         return [
             new TwigFunction('elfinder_tinymce_init', [$this, 'tinymce'], $options),
             new TwigFunction('elfinder_tinymce_init4', [$this, 'tinymce4'], $options),
+            new TwigFunction('elfinder_tinymce_init4', [$this, 'tinymce5'], $options),
             new TwigFunction('elfinder_summernote_init', [$this, 'summernote'], $options),
         ];
     }
 
     /**
-     * @param string $instance
-     * @param array  $parameters
-     *
      * @return mixed
      *
      * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
      */
-    public function tinymce($instance = 'default', $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0'])
+    public function tinymce(string $instance = 'default', array $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0'])
     {
         if (!is_string($instance)) {
             throw new RuntimeError('The function can be applied to strings only.');
@@ -65,21 +61,12 @@ class FMElfinderExtension extends AbstractExtension
     }
 
     /**
-     * @param string $instance
-     * @param array  $parameters
-     *
-     * @return mixed
-     *
      * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
      */
-    public function tinymce4($instance = 'default', $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0'])
+    public function tinymce4(string $instance = 'default', array $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0']): string
     {
-        if (!is_string($instance)) {
-            throw new RuntimeError('The function can be applied to strings only.');
-        }
-
         return $this->twig->render(
             '@FMElfinder/Elfinder/helper/_tinymce4.html.twig',
             [
@@ -91,23 +78,22 @@ class FMElfinderExtension extends AbstractExtension
         );
     }
 
+    public function tinymce5(string $instance = 'default'): string
+    {
+        return $this->twig->render(
+            '@FMElfinder/Elfinder/helper/_tinymce5.html.twig', [
+                'instance' => $instance,
+            ]
+        );
+    }
+
     /**
-     * @param string $instance
-     * @param string $selector
-     * @param array  $parameters
-     *
-     * @return mixed
-     *
      * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
      */
-    public function summernote($instance = 'default', $selector = '.summenote', $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0'])
+    public function summernote(string $instance = 'default', string $selector = '.summenote', array $parameters = ['width' => 900, 'height' => 450, 'title' => 'elFinder 2.0']): string
     {
-        if (!is_string($instance)) {
-            throw new RuntimeError('The function can be applied to strings only.');
-        }
-
         return $this->twig->render(
             '@FMElfinder/Elfinder/helper/_summernote.html.twig',
             [
@@ -125,7 +111,7 @@ class FMElfinderExtension extends AbstractExtension
      *
      * @see Twig_ExtensionInterface::getName()
      */
-    public function getName()
+    public function getName(): string
     {
         return 'fm_elfinder_init';
     }
